Add ability to edit posts from the edit link.
The update API endpoint returns the new post object.

It's not the most elegant solution.
The react component updates client side to force re-render.

I made a bad choice in design with the NewPost component.
It is used for both edits and new posts. This results in quite the nasty
confusion with notifying the parents.
(in this case, either a Post component or the App component itself)